Remove `generate` CLI root-command in favor of promoting its sub-commands to root-commands
