+++ src/lib.rs
@@ -3,7 +3,6 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 pub mod commands;
-pub mod generate;
 pub mod options;
 
 pub(crate) mod analyzer;
